+++ src/client/js/stack.js
@@ -1,0 +1,7 @@
+/*
+ * pwix:modal/src/client/js/stack.js
+ */
+
+import { mdStack } from '../classes/md_stack.class.js';
+
+Modal._stack = new mdStack();